+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.6.2-SNAPSHOT]
 
+- Form buttonbar alignment is now working correctly
 - UIComponent.currentWindow property
 - openModal/openWindow defaults to currentWindow as owner (https://github.com/edvin/tornadofx/issues/246)
 - Accordion.fold has `expanded` parameter